Abstract mSupply client
